% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD8T.R
\docType{data}
\name{CD8T}
\alias{CD8T}
\title{Dataset: CD8+ T cell gene expression data}
\format{
A data frame with 9369 rows (cells) and 1000 columns (genes).
}
\source{
Suo et al., Science (2022).
}
\usage{
data(CD8T)
}
\description{
The CD8T dataset provides the gene expression data of fetal CD8+ T cells
obtained in a single-cell RNA-seq experiment.
}
\references{
Suo, C., Dann, E., Goh, I., Jardine, L., Kleshchevnikov, V., Park, J.-E., Botting, R. A., et al.
"Mapping the developing human immune system across organs." Science 376(6597), eabo0510 (2022).
}
\keyword{datasets}
