% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covercorr_internal.R
\name{split_rectangles}
\alias{split_rectangles}
\title{Split rectangles by wrapping them around edges of \eqn{[0,1]^d}}
\usage{
split_rectangles(zmin, zmax)
}
\arguments{
\item{zmin}{n x d matrix of bottom-left coordinates, one row per rectangle}

\item{zmax}{n x d matrix of top-right coordinates, one row per rectangle}
}
\value{
a list of zmin and zmax, describing the bottom-left and top-right
coordinates of splitted rectangles
}
\description{
Split rectangles by wrapping them around edges of \eqn{[0,1]^d}
}
\details{
This is a wrapper of the C_split_rectangles function implemented in C
}
