% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covercorr_internal.R
\name{variance_formula}
\alias{variance_formula}
\title{Variance of the the excess vacancy}
\usage{
variance_formula(n, d)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension \eqn{(X, Y)}}
}
\value{
variance formula in paper
}
\description{
Exact formula for \eqn{n} times the variance of the excess vacancy.
For independent \eqn{X} and \eqn{Y}, the variance of the coverage correlation
coefficient is obtained by dividing the returned value by \eqn{n(1 - e^{-1})^2}.
check \href{https://arxiv.org/abs/2508.06402}{the arXiv preprint} for more details
}
