% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD8T.R
\docType{data}
\name{CD8T}
\alias{CD8T}
\title{Dataset: CD8+T cell gene expression data}
\format{
Data are provided in
raw sequencing counts as a data frame with 9369 rows (cells) and 1000 
columns (genes).
}
\source{
The original data were part of the published data by Suo and 
colleagues in Science (2022). Here we use only cells that were annotated 
to be CD8+T and the top 1000 highly variable genes.
}
\usage{
CD8T
}
\description{
The CD8T dataset provides the gene expression data of fetal CD8+T cells
obtained in single-cell RNA sequencing experiment.
}
\references{
Suo, Chenqu, Emma Dann, Issac Goh, Laura Jardine, Vitalii Kleshchevnikov, 
Jong-Eun Park, Rachel A. Botting et al. 
"Mapping the developing human immune system across organs." 
Science 376, no. 6597 (2022): eabo0510.
}
\keyword{datasets}
