% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{covered_volume_mc}
\alias{covered_volume_mc}
\title{Total volume of union of rectangles using Monte Carlo integration}
\usage{
covered_volume_mc(zmin, zmax, M)
}
\arguments{
\item{zmin}{n x d matrix of bottomleft coordinates, one row per rectangle}

\item{zmax}{n x d matrix of topright coordinates, one row per rectangle}

\item{M}{number of Monte Carlo integration sample points}
}
\value{
a list of the estimated volume of the union and its standard error
}
\description{
Total volume of union of rectangles using Monte Carlo integration
}
\details{
This is a wrapper of the C_covered_volume_mc function in C
}
