% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{covered_volume_partitioned}
\alias{covered_volume_partitioned}
\title{Total volume of union of rectangles using volume hashing}
\usage{
covered_volume_partitioned(zmin, zmax)
}
\arguments{
\item{zmin}{n x d matrix of bottomleft coordinates, one row per rectangle}

\item{zmax}{n x d matrix of topright coordinates, one row per rectangle}
}
\value{
a numeric value of the volume of the union
}
\description{
Total volume of union of rectangles using volume hashing
}
\details{
This is a wrapper of the C_covered_volume_partitioned function in C
}
