% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{MK_rank}
\alias{MK_rank}
\title{Compute Monge-Kantorovich ranks}
\usage{
MK_rank(X, U)
}
\arguments{
\item{X}{data points}

\item{U}{reference points}
}
\description{
Compute Monge-Kantorovich ranks
}
