% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{split_rectangles}
\alias{split_rectangles}
\title{Split axis-aligned rectangles mod [0,1]^d
Given lower and upper corners of axis-aligned rectangles in R^d, this 
function generates the image of these rectangles modulo [0,1]^d (i.e. 
viewing [0,1]^d with periodic boundaries)}
\usage{
split_rectangles(zmin, zmax)
}
\arguments{
\item{zmin}{A numeric matrix of size \eqn{n \times d} containing the lower
bounds (one rectangle per row) before shifting.}

\item{zmax}{A numeric matrix of size \eqn{n \times d} containing the upper
bounds (one rectangle per row) before shifting. Must satisfy
\code{zmin < zmax} elementwise.}
}
\value{
A list with two numeric matrices:
\describe{
  \item{\code{zmin_splitted}}{Matrix of lower bounds of the non-empty
  intersections with \eqn{[0,1]^d}.}
  \item{\code{zmax_splitted}}{Matrix of upper bounds of the non-empty
  intersections with \eqn{[0,1]^d}.}
}
}
\description{
Split axis-aligned rectangles mod [0,1]^d
Given lower and upper corners of axis-aligned rectangles in R^d, this 
function generates the image of these rectangles modulo [0,1]^d (i.e. 
viewing [0,1]^d with periodic boundaries)
}
