% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{split_rectangles}
\alias{split_rectangles}
\title{Split rectangels by wrapping them around edges of [0,1]^d}
\usage{
split_rectangles(zmin, zmax)
}
\arguments{
\item{zmin}{n x d matrix of bottomleft coordinates, one row per rectangle}

\item{zmax}{n x d matrix of topright coordinates, one row per rectangle}
}
\value{
a list of zmin and zmax, describing the bottomleft and topright 
coordinates of splitted rectangles
}
\description{
Split rectangels by wrapping them around edges of [0,1]^d
}
\details{
This is a wrapper of the C_split_rectangles function implemented in C
}
