% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_correlation.R
\name{variance_formula}
\alias{variance_formula}
\title{exact formula for variance of test stat}
\usage{
variance_formula(n, d)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}
}
\description{
exact formula for variance of test stat
}
